requirements.txt
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-CORS==3.0.10
Flask-Caching==2.3.0
redis==4.0.2
gevent==21.12.0
pymongo==4.1.1
bcrypt==3.2.0
psutil==6.0.0
python-dotenv==0.19.2
